requirements.txt
tqdm
openai
python-dotenv
numpy
networkx
tiktoken
pyecharts
wikipedia
tenacity
nltk
jieba
plotly
pandas
gradio>=5.44.1
kaleido
pyyaml
langcodes
requests
fastapi
trafilatura

leidenalg
igraph
python-louvain

# For visualization
matplotlib
